Move getDocType into route
