+++ web/app/components/document/index.hbs
@@ -10,6 +10,7 @@
       <Document::Sidebar
         @profile={{this.authenticatedUser.info}}
         @document={{@document}}
+        @docType={{@docType}}
         @deleteDraft={{perform this.deleteDraft}}
         @isCollapsed={{this.sidebarIsCollapsed}}
         @toggleCollapsed={{this.toggleSidebarCollapsedState}}
